fix sequence quantities
Fixed pu issue for the zero-sequence current. It also looks like the zero-sequence current
was being calculated in terms of magnitudes which is incorrect, fixed to use the complex values.
Also added positive- and negative-sequence currents, these are needed for admittance protection
and negative/zero-sequence overcurrent protection (used by CERTS). In case the sequence magnitudes are
needed, I added these to the results structure  along with the complex values.

Also did a bit of bikeshedding, I changed fault_studies to fault_study